chore: move middleware to middleware package
