+++ pkg/http/client.go
@@ -7,8 +7,8 @@ import (
 	"net/http"
 	"time"
 
+	"github.com/CloudLearnersOrg/golib/pkg/logger"
 	"github.com/gin-gonic/gin"
-	"github.com/google/uuid"
 )
 
 // Client wraps http.Client with tracing and logging capabilities
@@ -39,28 +39,6 @@ func (c *Client) OutgoingRequest(ctx *gin.Context, method, url string, body io.R
 	}
 
 	return c.Client.Do(req)
-}
-
-// Middleware returns a Gin middleware for incoming request logging
-func Middleware() gin.HandlerFunc {
-	return func(c *gin.Context) {
-		start := time.Now()
-
-		// Get or generate trace ID
-		traceID := c.GetHeader("X-Trace-ID")
-		if traceID == "" {
-			traceID = uuid.New().String()
-		}
-
-		// Set trace ID in context for further use
-		c.Set("X-Trace-ID", traceID)
-		c.Header("X-Trace-ID", traceID)
-
-		// Process request
-		c.Next()
-
-		incoming(c, traceID, time.Since(start))
-	}
 }
 
 type roundTripperFunc func(*http.Request) (*http.Response, error)
@@ -97,25 +75,6 @@ func extractTraceID(ctx context.Context) string {
 	return ""
 }
 
-func incoming(c *gin.Context, traceID string, duration time.Duration) {
-	attrs := []any{
-		"trace_id", traceID,
-		"http.response.status_code", c.Writer.Status(),
-		"http.request.method", c.Request.Method,
-		"http.route", c.Request.URL.Path,
-		"server.address", c.Request.Host,
-		"http.response.latency", duration.String(),
-	}
-
-	log(c.Writer.Status(), "incoming request", attrs...)
-	if c.Writer.Status() >= 400 {
-		slog.Error("incoming request failed", attrs...)
-		return
-	}
-
-	slog.Info("incoming request completed", attrs...)
-}
-
 func outgoing(req *http.Request, err error, resp *http.Response, traceID string, duration time.Duration) (*http.Response, error) {
 	attrs := []any{
 		"trace_id", traceID,
@@ -131,7 +90,7 @@ func outgoing(req *http.Request, err error, resp *http.Response, traceID string,
 	}
 
 	attrs = append(attrs, "http.response.status_code", resp.StatusCode)
-	log(resp.StatusCode, "outgoing request", attrs...)
+	logger.LogFilteredStatusCode(resp.StatusCode, "outgoing request", attrs...)
 	if resp.StatusCode >= 400 {
 		slog.Error("outgoing request failed", attrs...)
 		return resp, err
@@ -139,15 +98,4 @@ func outgoing(req *http.Request, err error, resp *http.Response, traceID string,
 
 	slog.Info("outgoing request completed", attrs...)
 	return resp, nil
-}
-
-func log(status int, prefix string, attrs ...any) {
-	switch {
-	case status >= 500:
-		slog.Error(prefix+" failed", attrs...)
-	case status >= 400:
-		slog.Warn(prefix+" warning", attrs...)
-	default:
-		slog.Info(prefix+" completed", attrs...)
-	}
 }
