+++ pkg/http/client_test.go
@@ -7,6 +7,7 @@ import (
 	"strings"
 	"testing"
 
+	"github.com/CloudLearnersOrg/golib/pkg/middlewares/logger"
 	"github.com/gin-gonic/gin"
 	"github.com/google/uuid"
 	"github.com/stretchr/testify/assert"
@@ -72,48 +73,6 @@ func TestOutgoingPostRequest(t *testing.T) {
 	assert.Equal(t, http.StatusOK, resp.StatusCode)
 }
 
-func TestMiddlewareWithoutTraceID(t *testing.T) {
-	// Given
-	gin.SetMode(gin.TestMode)
-	router := gin.New()
-	router.Use(Middleware())
-	router.GET("/test", func(c *gin.Context) {
-		c.Status(http.StatusOK)
-	})
-
-	resp := httptest.NewRecorder()
-	req := httptest.NewRequest(http.MethodGet, "/test", nil)
-
-	// When
-	router.ServeHTTP(resp, req)
-
-	// Then
-	assert.Equal(t, http.StatusOK, resp.Code)
-	assert.NotEmpty(t, resp.Header().Get("X-Trace-ID"))
-}
-
-func TestMiddlewareWithExistingTraceID(t *testing.T) {
-	// Given
-	gin.SetMode(gin.TestMode)
-	router := gin.New()
-	router.Use(Middleware())
-	router.GET("/test", func(c *gin.Context) {
-		c.Status(http.StatusOK)
-	})
-
-	resp := httptest.NewRecorder()
-	req := httptest.NewRequest(http.MethodGet, "/test", nil)
-	existingTraceID := uuid.New().String()
-	req.Header.Set("X-Trace-ID", existingTraceID)
-
-	// When
-	router.ServeHTTP(resp, req)
-
-	// Then
-	assert.Equal(t, http.StatusOK, resp.Code)
-	assert.Equal(t, existingTraceID, resp.Header().Get("X-Trace-ID"))
-}
-
 func TestTraceIDPropagationThroughChain(t *testing.T) {
 	// Given
 	downstream := httptest.NewServer(http.HandlerFunc(func(body http.ResponseWriter, req *http.Request) {
@@ -125,7 +84,7 @@ func TestTraceIDPropagationThroughChain(t *testing.T) {
 
 	gin.SetMode(gin.TestMode)
 	router := gin.New()
-	router.Use(Middleware())
+	router.Use(logger.Middleware())
 	client := NewClient(nil)
 
 	router.GET("/test", func(c *gin.Context) {
